Sync with upstream (#1)
* Add process Idle status

Currently, I is also a valid status for a process as shown in the
following example. This patch adds the Idle status to the list of
valid status.

me@labo:/proc$ cat /proc/75/stat
75 (nvme-reset-wq) I 2 0 0 0 -1 69238880 0 0 0 0 0 0 0 0 0 -20 1 0 \
133 0 0 4294967295 0 0 0 0 0 0 0 2147483647 0 0 0 0 17 2 0 0 0 0 0 \
0 0 0 0 0 0 0 0
me@labo:/proc$ cat /proc/75/status
Name:   nvme-reset-wq
Umask:  0000
State:  I (idle)
Tgid:   75
*snip*

* Fix some typos

* Remove deprecated code

ioutil is deprecated: As of Go 1.16, use os.ReadFile.

* Fix some linting issues

go-simple
nonamedreturns

* Fix error when splitting stat string only by space

observe:
me@labo:/proc$ cat /proc/79/stat
79 (DWC Notificatio) I 2 0 0 0 -1 69238880 0 0 0 0 0 0 0 0 0 -20 1 0 \
156 0 0 4294967295 0 0 0 0 0 0 0 2147483647 0 0 0 0 17 2 0 0 0 0 0 0 \
0 0 0 0 0 0 0

When splitting simply by space, the 2nd field is "DWC" and the 3rd is
"Notificatio".  This is not what we want. We want the 3rd field to be
an integer for the parent process ID. Otherwise the
strconv.Atoi(stat[3]) fails with a parse error on I. Hence we need an
exception on the splits if the space is encapsulated by parentheses.

* Remove unecessary var assignments

Also elaborate returned error messages to simplify debugging.
They now indicate in which function and when the error occured.

* [FIX] use os package

* [FIX] pgrep - use os package

---------

Co-authored-by: EricPark <69110284+Eric011025@users.noreply.github.com>
Co-authored-by: Eric011025 <ericpark011025@gmail.com>